🎨: #38 뷰모델을 나누기, 코드 스타일링
- ViewModel에서 리스트들로 데이터를 가져오는 코드들을 HostedViewModel과 InvitedViewModel로 나눠서 구현
- getUserUid() 함수의 코드 스타일 변경
